Restrukturalizace služeb na rozhraní a implementace, přidání null checků do relevantních metod implementací

+++ Project/DP_SUM0023/DP_SUM0023/Data/Services/EntityFramework/CompanyServiceEF.cs
@@ -1,7 +1,8 @@
-﻿using DP_SUM0023.Data.Models;
+﻿using DP_SUM0023.Data.Services.Interfaces;
+using DP_SUM0023.Data.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace DP_SUM0023.Data.Services
+namespace DP_SUM0023.Data.Services.EntityFramework
 {
     public class CompanyServiceEF : ICompanyService
     {
@@ -23,18 +24,27 @@ namespace DP_SUM0023.Data.Services
 
         public async Task CreateAsync(Company instanceToCreate)
         {
+            if (instanceToCreate == null)
+                return;
+
             await dbContext.Company.AddAsync(instanceToCreate);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task UpdateAsync(Company instanceToUpdate)
         {
+            if (instanceToUpdate == null)
+                return;
+
             dbContext.Company.Update(instanceToUpdate);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task RemoveAsync(Company instanceToRemove)
         {
+            if (instanceToRemove == null)
+                return;
+
             dbContext.Company.Remove(instanceToRemove);
             await dbContext.SaveChangesAsync();
         }
