+++ Project/DP_SUM0023/DP_SUM0023/Data/Services/EntityFramework/ProcessServiceEF.cs
@@ -1,7 +1,8 @@
-﻿using DP_SUM0023.Data.Models;
+﻿using DP_SUM0023.Data.Services.Interfaces;
+using DP_SUM0023.Data.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace DP_SUM0023.Data.Services
+namespace DP_SUM0023.Data.Services.EntityFramework
 {
     public class ProcessServiceEF : IProcessService
     {
@@ -23,18 +24,27 @@ namespace DP_SUM0023.Data.Services
 
         public async Task CreateAsync(Process instanceToCreate)
         {
+            if (instanceToCreate == null)
+                return;
+
             await dbContext.Process.AddAsync(instanceToCreate);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task UpdateAsync(Process instanceToUpdate)
         {
+            if (instanceToUpdate == null)
+                return;
+
             dbContext.Process.Update(instanceToUpdate);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task RemoveAsync(Process instanceToRemove)
         {
+            if (instanceToRemove == null)
+                return;
+
             dbContext.Process.Remove(instanceToRemove);
             await dbContext.SaveChangesAsync();
         }
