+++ Project/DP_SUM0023/DP_SUM0023/Data/Services/EntityFramework/ProjectServiceEF.cs
@@ -1,7 +1,8 @@
-﻿using DP_SUM0023.Data.Models;
+﻿using DP_SUM0023.Data.Services.Interfaces;
+using DP_SUM0023.Data.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace DP_SUM0023.Data.Services
+namespace DP_SUM0023.Data.Services.EntityFramework
 {
     public class ProjectServiceEF : IProjectService
     {
@@ -23,24 +24,36 @@ namespace DP_SUM0023.Data.Services
 
         public async Task CreateAsync(Project instanceToCreate)
         {
+            if (instanceToCreate == null)
+                return;
+
             await dbContext.Project.AddAsync(instanceToCreate);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task UpdateAsync(Project instanceToUpdate)
         {
+            if (instanceToUpdate == null)
+                return;
+
             dbContext.Project.Update(instanceToUpdate);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task RemoveAsync(Project instanceToRemove)
         {
+            if (instanceToRemove == null)
+                return;
+
             dbContext.Project.Remove(instanceToRemove);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task<List<Project>> GetProjectsNotAssignedToAccount(UserAccount account)
         {
+            if (account == null)
+                return new List<Project>();
+
             var assignedProjects = account.AssignedProjects;
             var allProjects = await GetAllAsync();
             return allProjects.Where(p => !assignedProjects.Contains(p)).ToList();
